requirements.txt
flask==1.0.2
requests==2.21.0
google-cloud-bigquery==1.9.0
google-cloud-firestore==0.31.0
selenium==3.141.0 